Order

+++ src/main/java/za/ac/cput/factory/OrderFactory.java
@@ -11,7 +11,7 @@ import za.ac.cput.util.Helper;
 import java.util.Date;
 
 public class OrderFactory {
-    public static Order createOrder(String orderID, Date orderDate, int totalAmount, String status){
+    public static Order createOrder(String orderID, Date orderDate, double totalAmount, String status){
         if(Helper.isNullOrEmpty(orderID) || orderDate == null || Helper.isNegative(totalAmount) ||Helper.isNullOrEmpty(status)){
             return null;
         }
@@ -22,7 +22,7 @@ public class OrderFactory {
                 .setStatus(status)
                 .build();
     }
-    public static Order createOrder(Date orderDate, int totalAmount, String status) {
+    public static Order createOrder(Date orderDate, double totalAmount, String status) {
         if (orderDate == null || Helper.isNegative(totalAmount) || Helper.isNullOrEmpty(status)) {
             return null;
         }
